requirements.txt
streamlit==1.35.0
pandas
openai==2.8.1
sqlite-utils
python-dotenv
altair
